Toubleshooting de useDeleteProject

+++ hooks/useDeleteProject.ts
@@ -1,13 +1,20 @@
+'use client'
+
 import { useState } from 'react'
-import Router from 'next/router'
+import { useRouter } from 'next/navigation'
 
 export default function useDeleteProject() {
 
     const [ isLoading, setIsLoading ] = useState<boolean>(false)
     const [ isSuccess, setIsSuccess ] = useState<boolean>(false)
+    const Router = useRouter()
 
     async function deleteProject(_id: string) {
         try {
+            if(!window.localStorage.getItem('token')) {
+                Router.push('/connexion')
+                throw new Error('Veuillez vous connecter')
+            }
             setIsLoading(true)
             const responseJSON = await fetch(`${process.env.NEXT_PUBLIC_SERVER_URL}/api/project/deleteProject`, {
             method: 'DELETE',
@@ -16,15 +23,18 @@ export default function useDeleteProject() {
             },            
             body: JSON.stringify({_id: _id })
         })
-        setIsLoading(false)
-        if(!responseJSON.ok) { throw new Error('Erreur de la requête http')}
-
+        
         const response = await responseJSON.json()
-        console.log('response',response)
-        if(response.status === 403) {
-            alert("L'authentification est nécessaire pour supprimer un projet")
-            Router.push('/connexion')
+        console.log('responseJSON.status',responseJSON.status)
+        
+        if(responseJSON.status === 403 || 401) {
+                Router.push('/connexion')
+                throw new Error(response.message)
+            }
+        if(!responseJSON.ok) {
+            throw new Error('Contacter votre administrateur')
         }
+        setIsLoading(false)     
         setIsSuccess(true)
         }
         catch (error) {
